requirements.txt
arxiv==0.5.3
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
feedparser==6.0.2
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
numpy==1.19.3
pandas==1.1.4
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMuPDF==1.18.3
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
rsa==4.6
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.0.1
uritemplate==3.0.1
urllib3==1.26.1
